# swagger-server/api/swagger.yaml
openapi: 3.0.0
info:
  title: User API
  version: 1.0.0
  description: API para gestionar usuarios
servers:
  - url: http://localhost:3500/api
    description: Servidor local

paths:
  /user/register:
    post:
      summary: Registra un nuevo usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
                - phones
              properties:
                name:
                  type: string
                  example: Juan Rodriguez
                email:
                  type: string
                  format: email
                  example: juan@rodriguez.org
                password:
                  type: string
                  example: '@Diego123'
                phones:
                  type: array
                  items:
                    type: object
                    properties:
                      number:
                        type: string
                        example: 1234567
                      citycode:
                        type: string
                        example: 1
                      contrycode:
                        type: string
                        example: 57
      responses:
        '201':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User created successfully!
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Error en los datos de entrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid email
  /user/all:
    get:
      summary: Obtiene todos los usuarios
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Acceso denegado

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: hashed string
        phones:
          type: array
          items:
            type: object
            properties:
              number:
                type: string
              citycode:
                type: string
              contrycode:
                type: string
        token:
          type: string
        isactive:
          type: boolean
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time